[ADD]: user view and admin view, more html, repositories, dtos

+++ timetracking/src/main/java/com/nunegal/timetracking/entity/Department.java
@@ -9,7 +9,7 @@ public class Department {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="iddepartment")
-    private Long id;
+    private int id;
 
     private String name;
 
@@ -17,13 +17,13 @@ public class Department {
 
     public Department() {}
 
-    public Department(Long id, String name, String description) {
+    public Department(int id, String name, String description) {
         this.id = id;
         this.name = name;
         this.description = description;
     }
-    public Long getId() {return id;}
-    public void setId(Long id) {this.id = id;}
+    public int getId() {return id;}
+    public void setId(int id) {this.id = id;}
     public String getName() {return name;}
     public void setName(String name) {this.name = name;}
     public String getDescription() {return description;}
