+++ timetracking/src/main/java/com/nunegal/timetracking/entity/Rol.java
@@ -9,21 +9,21 @@ public class Rol {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="idrol")
-    private Long id;
+    private int id;
 
     private String type;
 
     private String description;
 
     public Rol() {}
-    public Rol(Long id, String type, String description) {
+    public Rol(int id, String type, String description) {
         this.id = id;
         this.type = type;
         this.description = description;
     }
 
-    public Long getId() {return id;}
-    public void setId(Long id) {this.id = id;}
+    public int getId() {return id;}
+    public void setId(int id) {this.id = id;}
     public String getType() {return type;}
     public void setType(String type) {this.type = type;}
     public String getDescription() {return description;}
